moved letters generation into a dedicated module and actually switched from JS to TS
