+++ src/html/play.html
@@ -7,7 +7,7 @@
         <title>Words hunter</title>
         <link href="https://fonts.googleapis.com/css2?family=Source+Code+Pro&display=swap" rel="stylesheet">
         <link rel=stylesheet href="styles.css">
-        <script src="words-hunter.js"></script>
+        <script type="module" src="words-hunter.js"></script>
     </head>
     <body class="vbox-nowrap">
         <div class="vbox-nowrap canvas">
