+++ tsconfig.json
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
     "outDir": "./site",
-    "allowJs": true,
     "target": "es6"
   },
   "include": ["./src/**/*"]
